requirements.txt
bcrypt==3.2.0
email-validator==1.1.2

Werkzeug==2.0.0
jinja2==3.0.3
Flask==2.1.0
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
pep8==1.7.1
psycopg2
python-dotenv==0.17.0
SQLAlchemy==1.4.5
WTForms==2.3.3
requests
tzdata

Flask-Caching
redis
gunicorn
sendgrid


